refactor: Move MCTS configs to mcts module
MCTS config variants (PriorSamplingConfig, DecoupledPUCTConfig) now live
in alpharat/mcts/config.py where they belong. BatchMetadata imports from
there instead of defining them inline.
