+++ alpharat/mcts/__init__.py
@@ -1,5 +1,10 @@
 """Monte Carlo Tree Search for simultaneous-move games."""
 
+from alpharat.mcts.config import (
+    DecoupledPUCTConfig,
+    MCTSConfig,
+    PriorSamplingConfig,
+)
 from alpharat.mcts.nash import (
     compute_nash_equilibrium,
     compute_nash_value,
@@ -10,9 +15,12 @@ from alpharat.mcts.search import MCTSSearch, SearchResult
 from alpharat.mcts.tree import MCTSTree
 
 __all__ = [
+    "DecoupledPUCTConfig",
+    "MCTSConfig",
     "MCTSNode",
     "MCTSSearch",
     "MCTSTree",
+    "PriorSamplingConfig",
     "SearchResult",
     "compute_nash_equilibrium",
     "compute_nash_value",
